[Refactor]: Validation and Creating Array Chunk
Placed and simplified validation process for this module compared to
the previous version. We will need to consider if the chunk value is
'NaN' and should be an integer value only. It should also be greater than
zero;

Creating the result array with the chunk is also simplified. Overall
this is a simplier way of getting things done.